svg/painting/parsing/marker-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing marker with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#MarkerProperty" rel="help"/><h:meta content="marker supports the full paint grammar 'none | marker-ref'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("marker", "none");

test_valid_value("marker", 'url("https://example.com/")');

  ]]></script></svg>
